+++ package.json
@@ -10,7 +10,9 @@
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "style": "lessc --source-map-inline ./src/less/root.less ./static/css/root.css",
+    "tsc": "node ./tools/tsc.js",
+    "less": "node ./tools/lessc.js",
+    "contributors": "node ./tools/contributors.js",
     "lint-ts": "tslint -c tslint.json -p tsconfig.json -e \"node_modules/**/*.ts\"",
     "lint-style": "stylelint ./src/less/*.less"
   },
