+++ src/less/dialogs.less
@@ -18,11 +18,10 @@
 
 .tokenDialog {
   position: absolute;
-  z-index: 2;
   top: 60px;
   width: 300px;
   padding: 10px;
-  z-index: 1000;
+  z-index: 100;
   background-color: var(--electron-background-2);
   border: 1px solid var(--electron-dark);
   color: #f3f3f3;
