+++ src/less/main.less
@@ -12,7 +12,8 @@
   src: local('Roboto'), local('Roboto-Light'), url(../fonts/roboto/Roboto-Light.ttf) format('woff2');
 }
 
-html, body {
+html,
+body {
   padding: 0;
   margin: 0;
   background: @electron-background-2;
@@ -33,7 +34,7 @@ html, body {
   }
 
   *::-webkit-scrollbar-thumb {
-    background: rgba(255, 255, 255, .4);
+    background: rgba(255, 255, 255, 0.4);
   }
 
   *::-webkit-scrollbar-corner {
@@ -51,7 +52,7 @@ html, body {
     background-color: #9feaf9;
     border: none;
     color: #1e2527;
-    padding: 0px 30px;
+    padding: 0 30px;
     text-align: center;
     text-decoration: none;
     border-radius: 4px;
@@ -128,5 +129,3 @@ html, body {
   text-align: center;
   font-size: 12px;
 }
-
-
