+++ src/less/mosaic.less
@@ -20,7 +20,7 @@
     width: 20px;
     margin: 5px;
     background-color: #314246;
-    border: .1rem solid #222b35;
+    border: 0.1rem solid #222b35;
   }
 
   button.mosaic-default-control:hover {
