+++ src/less/output.less
@@ -5,15 +5,15 @@
   font-family: @default-fonts;
   font-weight: 600;
   color: #f3f3f3;
-  background: rgba(0,0,0, .4);
+  background: rgba(0, 0, 0, 0.4);
   top: 10vh;
   padding: 0;
   padding-left: 10px;
   margin-top: 0;
   height: 0;
   overflow-y: scroll;
-  transition: height .2s;
-  transition: padding .2s;
+  transition: height 0.2s;
+  transition: padding 0.2s;
   -webkit-app-region: no-drag;
 
   &.showing {
