+++ src/less/root.less
@@ -3,6 +3,7 @@
 
 @import "resets.less";
 @import "main.less";
+@import "commands.less";
 @import "output.less";
 @import "dialogs.less";
 @import "mosaic-vendor.less";
