+++ src/less/settings.less
@@ -10,7 +10,11 @@
   user-select: none;
   cursor: initial;
 
-  h1, h2, h3, h4, h5 {
+  h1,
+  h2,
+  h3,
+  h4,
+  h5 {
     color: white;
     cursor: initial;
   }
@@ -74,7 +78,7 @@
     .contributor {
       width: 320px;
       height: 120px;
-      background: rgba(255, 255, 255, .1);
+      background: rgba(255, 255, 255, 0.1);
       padding: 10px;
       margin: 0 10px 10px 0;
       display: flex;
@@ -82,10 +86,11 @@
       cursor: pointer;
 
       &:hover {
-        background: rgba(255, 255, 255, .15);
+        background: rgba(255, 255, 255, 0.15);
       }
 
-      h5, p {
+      h5,
+      p {
         margin-bottom: 0;
       }
 
@@ -117,7 +122,7 @@
     li {
       padding-bottom: 5px;
       border-bottom: solid 1px rgba(255, 255, 255, 0.2);
-      color: rgba(255, 255, 255, .8);
+      color: rgba(255, 255, 255, 0.8);
 
       &:hover {
         padding-bottom: 5px;
