+++ static/index.html
@@ -6,13 +6,7 @@
   <title>Electron Fiddle</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="../node_modules/milligram/dist/milligram.min.css">
-  <link rel="stylesheet" href="./css/mosaic-vendor.css">
-  <link rel="stylesheet" href="./css/main.css">
-  <link rel="stylesheet" href="./css/colors.css">
-  <link rel="stylesheet" href="./css/commands.css">
-  <link rel="stylesheet" href="./css/mosaic.css">
-  <link rel="stylesheet" href="./css/dialogs.css">
-  <link rel="stylesheet" href="./css/settings.css">
+  <link rel="stylesheet" href="./css/root.css">
 </head>
 <body class="fiddle">
   <div id="app"></div>
